requirements.txt
tensorflow
numpy
chess
python-chess
pillow
cairosvg
tqdm
